# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Interactive Fiction Platform API
  description: OpenAPI 3.0 Specification for Interactive Fiction Platform (DAG Novel System)
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
paths:
...
  /api/admin/categories:
    post:
      summary: Admin create a category
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/writers/apply:
    post:
      summary: Reader applies to become a writer
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [pen_name, main_contact]
              properties:
                full_name:
                  type: string
                  description: "ชื่อ-นามสกุลจริง"
                pen_name:
                  type: string
                  description: "นามปากกา (บังคับ)"
                bio:
                  type: string
                  description: "คำแนะนำตัวสั้นๆ"
                email:
                  type: string
                  format: email
                  description: "อีเมลสำหรับติดต่อนักเขียน"
                main_contact:
                  type: string
                  description: "ช่องทางติดต่อหลัก (บังคับ)"
                other_links:
                  type: string
                  description: "ช่องทางติดต่อเพิ่มเติม"
                genres:
                  type: string
                  description: "JSON array string เช่น [\"romance\",\"fantasy\"]"
                avatar:
                  type: string
                  format: binary
                  description: "รูปโปรไฟล์นักเขียน (ไม่บังคับ)"
      responses:
        '200':
          description: Writer application submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/writers/me:
    get:
      summary: Get writer application status for current user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Application status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [none, pending, approved, rejected]
                    description: "none = ไม่เคยสมัคร"
                  writer_id:
                    type: integer
                  full_name:
                    type: string
                  pen_name:
                    type: string
                  email:
                    type: string
                  bio:
                    type: string
                  main_contact:
                    type: string
                  other_links:
                    type: string
                  category_ids:
                    type: array
                    items:
                      type: integer
                  avatar_url:
                    type: string
                  rejection_reason:
                    type: string
                  rejected_at:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/writers/me/profile:
    put:
      summary: Update writer profile (approved writers only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pen_name]
              properties:
                pen_name:
                  type: string
                  description: "นามปากกา (บังคับ ห้ามเว้นว่าง)"
                bio:
                  type: string
                avatar_url:
                  type: string
                  description: "URL รูปโปรไฟล์ — ได้จากการ Upload ผ่าน POST /upload/image ก่อน"
                contact_required:
                  type: string
                  description: "ช่องทางติดต่อหลัก"
                contact_optional:
                  type: string
                  description: "ช่องทางติดต่อเพิ่มเติม"
                category_ids:
                  type: array
                  items:
                    type: integer
                  description: "รายการ ID หมวดหมู่ที่นักเขียนสังกัด"
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/writers/requests:
    get:
      summary: Admin list all pending writer applications
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of pending writer requests
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/writers/approve:
    post:
      summary: Admin approve writer application
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: writer_id
          required: true
          schema:
            type: integer
          description: "รหัสคำขอนักเขียนที่ต้องการอนุมัติ"
      responses:
        '200':
          description: Writer application approved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/writers/reject:
    post:
      summary: Admin reject writer application
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: writer_id
          required: true
          schema:
            type: integer
          description: "รหัสคำขอนักเขียนที่ต้องการปฏิเสธ"
      responses:
        '200':
          description: Writer application rejected
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/admin/audit-logs:
    get:
      summary: List admin audit logs
      security:
        - BearerAuth: []
      parameters:
        - {in: query, name: actor_user_id, schema: {type: integer, minimum: 1}}
        - {in: query, name: action, schema: {type: string}}
        - {in: query, name: target_type, schema: {type: string}}
        - {in: query, name: target_id, schema: {type: integer, minimum: 1}}
        - {in: query, name: status, schema: {type: string}}
        - {in: query, name: date_from, schema: {type: string, format: date-time}}
        - {in: query, name: date_to, schema: {type: string, format: date-time}}
        - {in: query, name: page, schema: {type: integer, minimum: 1, default: 1}}
        - {in: query, name: limit, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
      responses:
        '200':
          description: Audit log page
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuditLogListResponse'
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /api/admin/novels:
    get:
      summary: Admin list novels by moderation status
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [all, published, suspended]
            default: all
        - in: query
          name: category_id
          schema:
            type: integer
            minimum: 1
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated novels for admin moderation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminNovelListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/admin/reports:
    get:
      summary: Admin list reports and appeals
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [all, pending, resolved, rejected, appeal_pending]
            default: all
        - in: query
          name: type
          schema:
            type: string
            enum: [all, report, appeal]
            default: all
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated reports and appeals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Resource Conflict Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    SuccessResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: integer
        message:
          type: string
        data:
          nullable: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: BAD_REQUEST
            message:
              type: string
              example: Invalid input payload
...
    ReportListResponse:
      type: object
      required: [reports, page, limit, total]
      properties:
        reports:
          type: array
          items:
            $ref: '#/components/schemas/ReportResponse'
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
    ReportResponse:
      type: object
      required: [report_id, user_id, novel_id, reason, status, created_at, novel_cover, novel_synopsis, author_pen_name, author_user_id]
      properties:
        report_id:
          type: integer
        user_id:
          type: integer
        username:
          type: string
        novel_id:
          type: integer
        novel_title:
          type: string
        reason:
          type: string
        report_type:
          type: string
          enum: [report, appeal]
        status:
          type: string
          enum: [pending, resolved, rejected, appeal_pending]
        created_at:
          type: string
          format: date-time
        novel_cover:
          type: string
        novel_synopsis:
          type: string
        author_pen_name:
          type: string
        author_user_id:
          type: integer
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
...
    AdminNovelListResponse:
      type: object
      required: [novels, page, limit, total]
      properties:
        novels:
          type: array
          items:
            $ref: '#/components/schemas/AdminNovel'
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
    AdminNovel:
      type: object
      required: [novel_id, title, status, is_published, is_completed, author_id]
      properties:
        novel_id:
          type: integer
        title:
          type: string
        captions:
          type: string
          nullable: true
        introduction:
          type: string
          nullable: true
        cover_image:
          type: string
          nullable: true
        status:
          type: string
          enum: [published, suspended]
        is_published:
          type: boolean
        is_completed:
          type: boolean
        author_id:
          type: integer
        author_name:
          type: string
        pen_name:
          type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
...
    AuditLog:
      type: object
      required: [log_id, actor_role, action, status, metadata, created_at]
      properties:
        log_id: {type: integer, format: int64}
        actor_user_id: {type: integer, nullable: true}
        actor_username: {type: string, nullable: true}
        actor_role: {type: string}
        action: {type: string}
        target_type: {type: string, nullable: true}
        target_id: {type: integer, format: int64, nullable: true}
        target_name: {type: string, nullable: true}
        status: {type: string}
        ip_address: {type: string, nullable: true}
        metadata: {type: object, additionalProperties: true}
        created_at: {type: string, format: date-time}
    AuditLogListResponse:
      type: object
      required: [items, page, limit, total]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/AuditLog'}
        page: {type: integer}
        limit: {type: integer}
        total: {type: integer}
...